requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
python-dotenv==1.0.1
aiohttp==3.9.5
langchain==0.1.12
langchain-community==0.0.29
langchain-anthropic==0.1.5
langchain-openai==0.1.5
langgraph==0.1.6
pydantic==2.7.0
asyncio==3.4.3
loguru==0.7.2
python-multipart==0.0.9
anthropic==0.23.0
openai==1.37.2
setuptools>=49.2.1 